moved files around. change sendmessage arg to string instead of channel
