+++ src/agni/server/AgniServer.java
@@ -3,6 +3,8 @@ package agni.server;
 import java.nio.channels.SocketChannel;
 import java.util.ArrayList;
 
+import agni.server.communication.ChannelList;
+
 public class AgniServer {
 
 	public static void main(String[] args) {
